function/requirements.txt
functions_framework==3.8.0
protobuf==3.20.1
google-cloud-storage==2.5.0
Flask==2.0.1
werkzeug==2.0.3
numpy==1.26.3
